requirements.txt
openai
flask
requests
python-dotenv
flask-cors
gunicorn
